+++ landingzones/plugins.py
@@ -104,7 +104,7 @@ class ProjectAppPlugin(ProjectAppPluginPoint):
                     'zone_uuid': zone.sodar_uuid,
                     'user_name': zone.user.username,
                     'user_uuid': str(zone.user.sodar_uuid),
-                    'assay_path': irods_backend.get_subdir(
+                    'assay_path': irods_backend.get_sub_path(
                         zone.assay, landing_zone=True
                     ),
                     'description': zone.description,